+++ sickrage/core/databases/__init__.py
@@ -219,6 +219,7 @@ class srDatabase(object):
 
             import sqlite3
             conn = sqlite3.connect(self.old_db_path)
+            conn.text_factory = lambda x: (x.decode('utf-8', 'ignore'))
 
             migrate_data = {}
             rename_old = False
@@ -249,8 +250,6 @@ class srDatabase(object):
                                 migrate_data[ml][p[0]] = [migrate_data[ml][p[0]]]
                             migrate_data[ml][p[0]].append(columns)
 
-                conn.close()
-
                 sickrage.srCore.srLogger.info('Getting data took %s', (time.time() - migrate_start))
 
                 if not self.db.opened:
@@ -277,6 +276,8 @@ class srDatabase(object):
                 rename_old = True
             except:
                 sickrage.srCore.srLogger.debug('Migration of %s database failed', self.name)
+            finally:
+                conn.close()
 
             # rename old database
             if rename_old: